UX-310 Fix global header inheritance in code editor, Add copy buttons, Adjust side nav link color
