+++ site/src/components/Heading.js
@@ -5,7 +5,17 @@ import { tokens } from '@sparkpost/design-tokens';
 import styled from 'styled-components';
 
 const StyledLink = styled.a`
+  display: block;
   text-decoration: none;
+  padding-top: ${tokens.spacing_400};
+  margin-bottom: ${tokens.spacing_400};
+  color: ${tokens.color_gray_900};
+
+  &:hover,
+  &:visited {
+    color: ${tokens.color_gray_900};
+  }
+
   span {
     opacity: 0;
     color: ${tokens.color_blue_700};
